+++ package.json
@@ -44,12 +44,10 @@
 		"@types/react": "^19.1.10",
 		"@types/react-dom": "^19.1.7",
 		"@vitejs/plugin-react": "^5.0.0",
-		"autoprefixer": "^10.4.21",
 		"depcheck": "^1.4.7",
 		"husky": "^9.1.7",
 		"jsdom": "^27.0.0",
 		"madge": "^8.0.0",
-		"postcss": "^8.5.6",
 		"rollup-plugin-visualizer": "^6.0.3",
 		"tailwindcss": "^4.1.13",
 		"terser": "^5.44.0",
